PM-3134: Extract in-memory store.
